+++ cli.js
@@ -4,6 +4,11 @@ const meow = require('meow')
 const open = require('react-dev-utils/openBrowser')
 const chalk = require('chalk')
 const ok = require('ok-cli')
+const remark = {
+  emoji: require('remark-emoji'),
+  unwrapImages: require('remark-unwrap-images')
+}
+const pkg = require('./package.json')
 
 const config = require('pkg-conf').sync('mdx-deck')
 
@@ -38,37 +43,52 @@ const getConfig = conf => {
             ].map(require.resolve)
           }
         },
-        require.resolve('./lib/loader.js'),
+        {
+          loader: require.resolve('./lib/loader.js'),
+          options: {
+            mdPlugins: [
+              remark.emoji,
+              remark.unwrapImages
+            ]
+          }
+        }
       ]
     }
   ]
-  conf.module.rules[1].include.push(
-    path.join(__dirname, './src')
-  )
 
   return conf
 }
 
 const cli = meow(`
-  ${chalk.magenta('[mdx-deck]')}
-
   ${chalk.gray('Usage')}
 
     $ ${chalk.magenta('mdx-deck deck.mdx')}
 
     $ ${chalk.magenta('mdx-deck build deck.mdx')}
 
+    $ ${chalk.magenta('mdx-deck pdf deck.mdx')}
+
   ${chalk.gray('Options')}
-
-    -p --port     Dev server port
-
-    --no-open     Prevent from opening in default browser
-
-    -d --out-dir  Output directory for exporting
 
     --title       Title for the HTML document
 
+    ${chalk.gray('Dev server options')}
+
+      -p --port     Dev server port
+      --no-open     Prevent from opening in default browser
+
+    ${chalk.gray('Build options')}
+
+      -d --out-dir  Output directory for exporting
+
+    ${chalk.gray('PDF options')}
+
+      --out-file    Filename for PDF export
+      --width       Width in pixels
+      --heigh       Height in pixels
+
 `, {
+  description: chalk.magenta('[mdx-deck] ') + chalk.gray(pkg.description),
   flags: {
     port: {
       type: 'string',
@@ -85,6 +105,9 @@ const cli = meow(`
     },
     title: {
       type: 'string'
+    },
+    outFile: {
+      type: 'string',
     }
   }
 })
@@ -95,24 +118,48 @@ const doc = file || cmd
 if (!doc) cli.showHelp(0)
 
 const opts = Object.assign({
-  entry: path.join(__dirname, './src/entry.js'),
+  entry: path.join(__dirname, './dist/entry.js'),
   dirname: path.dirname(path.resolve(doc)),
   globals: {
     DOC_FILENAME: JSON.stringify(path.resolve(doc))
   },
   config: getConfig,
   title: 'mdx-deck',
-  outDir: 'dist'
+  port: 8080,
+  outDir: 'dist',
+  outFile: 'presentation.pdf'
 }, config, cli.flags)
 
 opts.outDir = path.resolve(opts.outDir)
 
 switch (cmd) {
   case 'build':
-    log('exporting')
+    log('building')
     ok.build(opts)
       .then(res => {
         log('done')
+      })
+      .catch(err => {
+        log.error(err)
+        process.exit(1)
+      })
+    break
+  case 'pdf':
+    log('exporting to PDF')
+    const pdf = require('./lib/pdf')
+    ok(opts)
+      .then(({ server }) => {
+        log('rendering PDF')
+        pdf(opts)
+          .then(filename => {
+            server.close()
+            log('done', filename)
+            process.exit(0)
+          })
+          .catch(err => {
+            log.error(err)
+            process.exit(1)
+          })
       })
       .catch(err => {
         log.error(err)
@@ -125,7 +172,7 @@ switch (cmd) {
     ok(opts)
       .then(res => {
         const url = 'http://localhost:' + res.port
-        open(url)
+        if (opts.open) open(url)
         log('listening on', chalk.magenta(url))
       })
       .catch(err => {