+++ CHANGELOG.md
@@ -1,6 +1,117 @@
 
 # Changelog
 
+## v1.5.10 2018-08-05
+
+- Update overview mode styles
+- Add grid view mode
+- Update docs
+
+## v1.5.9 2018-08-05
+
+- Update docs
+
+## v1.5.8 2018-08-05
+
+- Add support for `components` and `Provider` in themes
+
+## v1.5.7 2018-08-05
+
+- Add more built-in themes
+
+## v1.5.6 2018-08-05
+
+- Add invisible buttons to left and right for use on mobile devices
+
+## v1.5.5 2018-08-05
+
+- Update docs
+
+## v1.5.4 2018-08-04
+
+- Add docs for syntax highlighting
+
+## v1.5.3 2018-08-04
+
+- Add overview mode to see multiple slides at once
+- Add default layouts for inverting colors and creating a split layout slide
+
+## v1.5.2 2018-08-04
+
+- Add default styles for tables
+
+## v1.5.1 2018-08-04
+
+- Use remark-unwrap-images plugin
+
+## v1.5.0 2018-08-04
+
+- Add syntax highlighting option for fenced code blocks
+
+## v1.4.4 2018-08-04
+
+- Fix for how Appear children display on slide change
+
+## v1.4.3 2018-08-04
+
+- Update build setup for smaller package
+- Adjust keyboard shortcuts
+
+## v1.4.2 2018-08-03
+
+- Update ok-cli for better HTML output
+
+## v1.4.1 2018-08-03
+
+- Update docs
+- Add `yellow` theme
+
+## v1.4.0 2018-08-03
+
+- Adds Appear component
+- Adds propTypes to components
+- Update README
+
+## v1.3.2 2018-08-02
+
+- Remove default `target="_blank"` from links
+- Move custom Provider component into app
+
+## v1.3.1 2018-08-02
+
+- Add speaker notes markdown syntax and component
+
+## v1.3.0 2018-08-02
+
+- Add presenter mode with preview of next slide and timer
+
+## v1.2.3 2018-08-01
+
+- Fix `history.pushState` hash
+
+## v1.2.2 2018-07-31
+
+- Update dev server for static file server
+
+## v1.2.1 2018-07-31
+
+- Merge custom components with defaults
+
+## v1.2.0 2018-07-31
+
+- Add PDF export to CLI
+
+## v1.1.3 2018-07-31
+
+- Add emoji support
+- Update `.npmignore`
+
+## v1.1.2 2018-07-31
+
+- Fix `--no-open` option
+- Add ability to ignore key events
+- Normalize newlines for cross-platform compatibility
+
 ## v1.1.1 2018-07-31
 
 - Fix for supporting markdown tables